Separate SV into different components
